// frontend/src/api/gen/schemas/eventRole.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: 0.1.0
 */

export type EventRole = (typeof EventRole)[keyof typeof EventRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EventRole = {
  Admin: "Admin",
  Mentor: "Mentor",
  Participant: "Participant",
  SidequestMaster: "SidequestMaster",
  Stakeholder: "Stakeholder",
} as const;