// frontend/src/api/gen/schemas/expertRatingLeaderboardEntry.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: 0.1.0
 */
import type { ExpertRatingLeaderboardEntryCategories } from "./expertRatingLeaderboardEntryCategories";

export interface ExpertRatingLeaderboardEntry {
  categories: ExpertRatingLeaderboardEntryCategories;
  rating: number;
  team_id: string;
}